requirements.txt
telethon
pytz
Pillow
